requirements.txt
asgiref==3.8.1
boto3==1.38.17
botocore==1.38.17
Brotli==1.1.0
click==8.2.0
da-vinci==0.5.0
dj-database-url==2.3.0
Django==5.2.1
django-storages==1.14.6
easy-thumbnails==2.10
ffmpeg-python==0.2.0
future==1.0.0
gunicorn==23.0.0
h11==0.16.0
jmespath==1.0.1
packaging==25.0
pillow==11.2.1
psycopg==3.2.8
psycopg-binary==3.2.8
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-magic==0.4.27
s3transfer==0.12.0
setuptools==78.1.1
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wheel==0.45.1
whitenoise==6.9.0
